Refactor storage: Remove ColdStorage trait and FileColdStorage
