+++ benches/cold_storage.rs
@@ -31,8 +31,7 @@ use gallifreydb::core::id::{NodeId, VersionId};
 use gallifreydb::core::interning::GLOBAL_INTERNER;
 use gallifreydb::core::property::PropertyMapBuilder;
 use gallifreydb::core::temporal::BiTemporalInterval;
-use gallifreydb::storage::cold_storage::{ColdStorage, CompressionAlgorithm};
-use gallifreydb::storage::redb_cold_storage::{RedbColdStorage, RedbConfig};
+use gallifreydb::storage::redb_cold_storage::{CompressionAlgorithm, RedbColdStorage, RedbConfig};
 use gallifreydb::storage::version::NodeVersion;
 use std::time::Instant;
 use tempfile::TempDir;
