+++ src/storage/redb_cold_storage.rs
@@ -22,7 +22,6 @@
 //!
 //! ```ignore
 //! use gallifreydb::storage::redb_cold_storage::{RedbColdStorage, RedbConfig};
-//! use gallifreydb::storage::ColdStorage;
 //!
 //! let config = RedbConfig::default();
 //! let storage = RedbColdStorage::new("data/cold.redb", config)?;
@@ -35,19 +34,18 @@
 //! ```
 
 use crate::core::id::VersionId;
-use crate::storage::cold_storage::{
-    AtomicColdStorageStats, ColdStorage, ColdStorageConfig, ColdStorageStats, CompressionAlgorithm,
-    decode_edge_version, decode_node_version, encode_edge_version, encode_node_version,
-};
 use crate::storage::version::{EdgeVersion, NodeVersion};
 use crate::storage::wal::LSN;
 use crate::utils::error::{Result, StorageError};
 use redb::{ReadableDatabase, ReadableTable};
 use std::path::{Path, PathBuf};
-use std::sync::atomic::Ordering;
+use std::sync::atomic::{AtomicU64, Ordering};
 
 #[cfg(feature = "config-toml")]
 use serde::{Deserialize, Serialize};
+
+#[cfg(test)]
+use std::sync::atomic::AtomicBool;
 
 // Table definitions with static lifetimes
 const NODE_VERSIONS_TABLE: redb::TableDefinition<'static, u64, &'static [u8]> =
@@ -61,10 +59,152 @@ const METADATA_TABLE: redb::TableDefinition<'static, &'static str, &'static [u8]
 const FLUSHED_LSN_KEY: &str = "flushed_lsn";
 
 // ============================================================================
+// Types moved from cold_storage.rs
+// ============================================================================
+
+/// Compression algorithm for cold storage.
+#[derive(Debug, Clone, Copy, PartialEq, Eq, Default)]
+#[cfg_attr(feature = "config-toml", derive(Serialize, Deserialize))]
+pub enum CompressionAlgorithm {
+    /// No compression (fastest, but uses more disk space)
+    None,
+    /// Zstd compression (ratio-optimized, default)
+    #[default]
+    Zstd,
+    /// LZ4-compatible fast compression using Zstd level 1
+    /// Provides similar speed characteristics to LZ4 with better compatibility
+    Fast,
+}
+
+impl CompressionAlgorithm {
+    /// Get the zstd compression level for this algorithm.
+    pub fn zstd_level(&self) -> Option<i32> {
+        match self {
+            CompressionAlgorithm::None => None,
+            CompressionAlgorithm::Zstd => Some(3), // Balanced ratio/speed
+            CompressionAlgorithm::Fast => Some(1), // Speed-optimized
+        }
+    }
+}
+
+/// Configuration for cold storage.
+#[derive(Debug, Clone)]
+#[cfg_attr(feature = "config-toml", derive(Serialize, Deserialize))]
+#[cfg_attr(feature = "config-toml", serde(default))]
+pub struct ColdStorageConfig {
+    /// Compression algorithm to use.
+    pub compression: CompressionAlgorithm,
+
+    /// Whether to sync writes to disk immediately.
+    /// When false, relies on OS buffer cache (faster but less durable).
+    pub sync_writes: bool,
+
+    /// Maximum number of versions to batch in a single write operation.
+    /// Higher values improve throughput but increase memory usage.
+    pub batch_size: usize,
+
+    /// Enable CRC32 checksums for data integrity verification.
+    pub enable_checksums: bool,
+}
+
+impl Default for ColdStorageConfig {
+    fn default() -> Self {
+        Self {
+            compression: CompressionAlgorithm::Zstd,
+            sync_writes: false,
+            batch_size: 1000,
+            enable_checksums: true,
+        }
+    }
+}
+
+/// Statistics for cold storage operations.
+#[derive(Debug, Clone, Default)]
+pub struct ColdStorageStats {
+    /// Total number of node versions stored.
+    pub node_versions_stored: u64,
+    /// Total number of edge versions stored.
+    pub edge_versions_stored: u64,
+    /// Total number of node version reads.
+    pub node_version_reads: u64,
+    /// Total number of edge version reads.
+    pub edge_version_reads: u64,
+    /// Total bytes written (before compression).
+    pub bytes_written_raw: u64,
+    /// Total bytes written (after compression).
+    pub bytes_written_compressed: u64,
+    /// Total bytes read (compressed size from storage).
+    pub bytes_read_compressed: u64,
+    /// Total bytes read (after decompression).
+    pub bytes_read_decompressed: u64,
+    /// Number of read errors.
+    pub read_errors: u64,
+    /// Number of write errors.
+    pub write_errors: u64,
+}
+
+impl ColdStorageStats {
+    /// Calculate the compression ratio (raw/compressed).
+    pub fn compression_ratio(&self) -> f64 {
+        if self.bytes_written_compressed == 0 {
+            1.0
+        } else {
+            self.bytes_written_raw as f64 / self.bytes_written_compressed as f64
+        }
+    }
+}
+
+/// Atomic statistics tracker for cold storage.
+#[derive(Debug, Default)]
+pub struct AtomicColdStorageStats {
+    /// Total number of node versions stored.
+    pub node_versions_stored: AtomicU64,
+    /// Total number of edge versions stored.
+    pub edge_versions_stored: AtomicU64,
+    /// Total number of node version reads.
+    pub node_version_reads: AtomicU64,
+    /// Total number of edge version reads.
+    pub edge_version_reads: AtomicU64,
+    /// Total bytes written (before compression).
+    pub bytes_written_raw: AtomicU64,
+    /// Total bytes written (after compression).
+    pub bytes_written_compressed: AtomicU64,
+    /// Total bytes read (compressed size from storage).
+    pub bytes_read_compressed: AtomicU64,
+    /// Total bytes read (after decompression).
+    pub bytes_read_decompressed: AtomicU64,
+    /// Number of read errors.
+    pub read_errors: AtomicU64,
+    /// Number of write errors.
+    pub write_errors: AtomicU64,
+}
+
+impl AtomicColdStorageStats {
+    /// Create a new atomic statistics tracker.
+    pub fn new() -> Self {
+        Self::default()
+    }
+
+    /// Create a snapshot of the current statistics.
+    pub fn snapshot(&self) -> ColdStorageStats {
+        ColdStorageStats {
+            node_versions_stored: self.node_versions_stored.load(Ordering::Relaxed),
+            edge_versions_stored: self.edge_versions_stored.load(Ordering::Relaxed),
+            node_version_reads: self.node_version_reads.load(Ordering::Relaxed),
+            edge_version_reads: self.edge_version_reads.load(Ordering::Relaxed),
+            bytes_written_raw: self.bytes_written_raw.load(Ordering::Relaxed),
+            bytes_written_compressed: self.bytes_written_compressed.load(Ordering::Relaxed),
+            bytes_read_compressed: self.bytes_read_compressed.load(Ordering::Relaxed),
+            bytes_read_decompressed: self.bytes_read_decompressed.load(Ordering::Relaxed),
+            read_errors: self.read_errors.load(Ordering::Relaxed),
+            write_errors: self.write_errors.load(Ordering::Relaxed),
+        }
+    }
+}
+
+// ============================================================================
 // Error Handling Helpers
 // ============================================================================
-// These helper functions consolidate error handling to reduce the number of
-// closures that LCOV counts as separate "functions" for coverage metrics.
 
 #[inline]
 fn map_io_error(context: &str) -> impl Fn(std::io::Error) -> crate::utils::error::Error + '_ {
@@ -170,9 +310,6 @@ impl RedbConfig {
 }
 
 /// Redb-based cold storage implementation.
-///
-/// This implementation uses Redb for durable, crash-safe storage of historical
-/// versions. It supports LSN tracking for WAL truncation coordination.
 pub struct RedbColdStorage {
     /// Path to the database file.
     path: PathBuf,
@@ -182,19 +319,15 @@ pub struct RedbColdStorage {
     config: RedbConfig,
     /// Statistics tracker.
     stats: AtomicColdStorageStats,
+    /// Fault injection flag for testing.
+    #[cfg(test)]
+    fail_writes: AtomicBool,
+    #[cfg(test)]
+    writes_attempted: AtomicBool,
 }
 
 impl RedbColdStorage {
     /// Create a new Redb cold storage at the given path.
-    ///
-    /// # Arguments
-    ///
-    /// * `path` - Path to the database file
-    /// * `config` - Configuration options
-    ///
-    /// # Returns
-    ///
-    /// Returns the storage instance or an error if database creation fails.
     pub fn new<P: AsRef<Path>>(path: P, config: RedbConfig) -> Result<Self> {
         let path = path.as_ref().to_path_buf();
 
@@ -232,6 +365,10 @@ impl RedbColdStorage {
             db,
             config,
             stats: AtomicColdStorageStats::new(),
+            #[cfg(test)]
+            fail_writes: AtomicBool::new(false),
+            #[cfg(test)]
+            writes_attempted: AtomicBool::new(false),
         })
     }
 
@@ -253,6 +390,30 @@ impl RedbColdStorage {
     /// Decompress data using the configured algorithm.
     fn decompress(&self, data: &[u8]) -> Result<Vec<u8>> {
         crate::storage::compression::decompress(data, &self.config.to_cold_storage_config())
+    }
+
+    /// Set fault injection flag for write operations (Test only).
+    #[cfg(test)]
+    pub fn set_fail_writes(&self, fail: bool) {
+        self.fail_writes.store(fail, Ordering::SeqCst);
+    }
+
+    /// Check if a write failure was injected and attempted (Test only).
+    #[cfg(test)]
+    pub fn was_write_attempted(&self) -> bool {
+        self.writes_attempted.load(Ordering::SeqCst)
+    }
+
+    /// Helper to check failure injection
+    fn check_fail_writes(&self) -> Result<()> {
+        #[cfg(test)]
+        {
+            self.writes_attempted.store(true, Ordering::SeqCst);
+            if self.fail_writes.load(Ordering::SeqCst) {
+                return Err(StorageError::io_error("Simulated write failure").into());
+            }
+        }
+        Ok(())
     }
 
     /// Get the flushed LSN from the metadata table.
@@ -286,10 +447,6 @@ impl RedbColdStorage {
     }
 
     /// Set the flushed LSN in the metadata table (internal helper).
-    ///
-    /// CRITICAL: Uses max(current_lsn, new_lsn) to prevent race conditions where
-    /// concurrent commits could overwrite a higher LSN with a lower LSN. This ensures
-    /// monotonic increase only, which is essential for WAL truncation safety.
     fn set_flushed_lsn_internal(
         table: &mut redb::Table<'_, &'static str, &'static [u8]>,
         lsn: LSN,
@@ -327,6 +484,413 @@ impl RedbColdStorage {
         Ok(())
     }
 
+    // ========================================================================
+    // Logic moved from ColdStorage trait impl
+    // ========================================================================
+
+    /// Store a node version.
+    pub fn store_node_version(&self, version: &NodeVersion) -> Result<()> {
+        self.check_fail_writes()?;
+
+        let encoded = encode_node_version(version);
+        let raw_size = encoded.len();
+        let compressed = self.compress(&encoded)?;
+        let compressed_size = compressed.len();
+
+        let write_txn = self
+            .db
+            .begin_write()
+            .map_err(map_transaction_error("Failed to begin write transaction"))?;
+
+        {
+            let mut table = write_txn.open_table(NODE_VERSIONS_TABLE).map_err(
+                |e| -> crate::utils::error::Error {
+                    StorageError::io_error(format!("Failed to open node_versions table: {}", e))
+                        .into()
+                },
+            )?;
+
+            table
+                .insert(version.id.as_u64(), compressed.as_slice())
+                .map_err(map_storage_error("Failed to store node version"))?;
+        }
+
+        write_txn
+            .commit()
+            .map_err(map_commit_error("Failed to commit"))?;
+
+        self.stats
+            .node_versions_stored
+            .fetch_add(1, Ordering::Relaxed);
+        self.stats
+            .bytes_written_raw
+            .fetch_add(raw_size as u64, Ordering::Relaxed);
+        self.stats
+            .bytes_written_compressed
+            .fetch_add(compressed_size as u64, Ordering::Relaxed);
+
+        Ok(())
+    }
+
+    /// Retrieve a node version.
+    pub fn get_node_version(&self, id: VersionId) -> Result<Option<NodeVersion>> {
+        self.stats
+            .node_version_reads
+            .fetch_add(1, Ordering::Relaxed);
+
+        let read_txn = self
+            .db
+            .begin_read()
+            .map_err(map_transaction_error("Failed to begin read transaction"))?;
+
+        let table = read_txn.open_table(NODE_VERSIONS_TABLE).map_err(
+            |e| -> crate::utils::error::Error {
+                StorageError::io_error(format!("Failed to open node_versions table: {}", e)).into()
+            },
+        )?;
+
+        match table.get(id.as_u64()) {
+            Ok(Some(value)) => {
+                let compressed: &[u8] = value.value();
+                self.stats
+                    .bytes_read_compressed
+                    .fetch_add(compressed.len() as u64, Ordering::Relaxed);
+
+                let decompressed = self.decompress(compressed)?;
+                self.stats
+                    .bytes_read_decompressed
+                    .fetch_add(decompressed.len() as u64, Ordering::Relaxed);
+
+                let version = decode_node_version(&decompressed)?;
+                Ok(Some(version))
+            }
+            Ok(None) => Ok(None),
+            Err(e) => {
+                Err(StorageError::io_error(format!("Failed to read node version: {}", e)).into())
+            }
+        }
+    }
+
+    /// Retrieve a batch of node versions.
+    pub fn get_node_versions_batch(&self, ids: &[VersionId]) -> Result<Vec<Option<NodeVersion>>> {
+        ids.iter().map(|id| self.get_node_version(*id)).collect()
+    }
+
+    /// Store an edge version.
+    pub fn store_edge_version(&self, version: &EdgeVersion) -> Result<()> {
+        self.check_fail_writes()?;
+
+        let encoded = encode_edge_version(version);
+        let raw_size = encoded.len();
+        let compressed = self.compress(&encoded)?;
+        let compressed_size = compressed.len();
+
+        let write_txn = self
+            .db
+            .begin_write()
+            .map_err(map_transaction_error("Failed to begin write transaction"))?;
+
+        {
+            let mut table = write_txn.open_table(EDGE_VERSIONS_TABLE).map_err(
+                |e| -> crate::utils::error::Error {
+                    StorageError::io_error(format!("Failed to open edge_versions table: {}", e))
+                        .into()
+                },
+            )?;
+
+            table
+                .insert(version.id.as_u64(), compressed.as_slice())
+                .map_err(map_storage_error("Failed to store edge version"))?;
+        }
+
+        write_txn
+            .commit()
+            .map_err(map_commit_error("Failed to commit"))?;
+
+        self.stats
+            .edge_versions_stored
+            .fetch_add(1, Ordering::Relaxed);
+        self.stats
+            .bytes_written_raw
+            .fetch_add(raw_size as u64, Ordering::Relaxed);
+        self.stats
+            .bytes_written_compressed
+            .fetch_add(compressed_size as u64, Ordering::Relaxed);
+
+        Ok(())
+    }
+
+    /// Retrieve an edge version.
+    pub fn get_edge_version(&self, id: VersionId) -> Result<Option<EdgeVersion>> {
+        self.stats
+            .edge_version_reads
+            .fetch_add(1, Ordering::Relaxed);
+
+        let read_txn = self
+            .db
+            .begin_read()
+            .map_err(map_transaction_error("Failed to begin read transaction"))?;
+
+        let table = read_txn.open_table(EDGE_VERSIONS_TABLE).map_err(
+            |e| -> crate::utils::error::Error {
+                StorageError::io_error(format!("Failed to open edge_versions table: {}", e)).into()
+            },
+        )?;
+
+        match table.get(id.as_u64()) {
+            Ok(Some(value)) => {
+                let compressed: &[u8] = value.value();
+                self.stats
+                    .bytes_read_compressed
+                    .fetch_add(compressed.len() as u64, Ordering::Relaxed);
+
+                let decompressed = self.decompress(compressed)?;
+                self.stats
+                    .bytes_read_decompressed
+                    .fetch_add(decompressed.len() as u64, Ordering::Relaxed);
+
+                let version = decode_edge_version(&decompressed)?;
+                Ok(Some(version))
+            }
+            Ok(None) => Ok(None),
+            Err(e) => {
+                Err(StorageError::io_error(format!("Failed to read edge version: {}", e)).into())
+            }
+        }
+    }
+
+    /// Retrieve a batch of edge versions.
+    pub fn get_edge_versions_batch(&self, ids: &[VersionId]) -> Result<Vec<Option<EdgeVersion>>> {
+        ids.iter().map(|id| self.get_edge_version(*id)).collect()
+    }
+
+    /// Check if a node version exists.
+    pub fn contains_node_version(&self, id: VersionId) -> Result<bool> {
+        let read_txn = self
+            .db
+            .begin_read()
+            .map_err(map_transaction_error("Failed to begin read transaction"))?;
+
+        let table = read_txn.open_table(NODE_VERSIONS_TABLE).map_err(
+            |e| -> crate::utils::error::Error {
+                StorageError::io_error(format!("Failed to open node_versions table: {}", e)).into()
+            },
+        )?;
+
+        match table.get(id.as_u64()) {
+            Ok(Some(_)) => Ok(true),
+            Ok(None) => Ok(false),
+            Err(e) => {
+                Err(StorageError::io_error(format!("Failed to check node version: {}", e)).into())
+            }
+        }
+    }
+
+    /// Check if an edge version exists.
+    pub fn contains_edge_version(&self, id: VersionId) -> Result<bool> {
+        let read_txn = self
+            .db
+            .begin_read()
+            .map_err(map_transaction_error("Failed to begin read transaction"))?;
+
+        let table = read_txn.open_table(EDGE_VERSIONS_TABLE).map_err(
+            |e| -> crate::utils::error::Error {
+                StorageError::io_error(format!("Failed to open edge_versions table: {}", e)).into()
+            },
+        )?;
+
+        match table.get(id.as_u64()) {
+            Ok(Some(_)) => Ok(true),
+            Ok(None) => Ok(false),
+            Err(e) => {
+                Err(StorageError::io_error(format!("Failed to check edge version: {}", e)).into())
+            }
+        }
+    }
+
+    /// Delete a node version.
+    pub fn delete_node_version(&self, id: VersionId) -> Result<bool> {
+        let write_txn = self
+            .db
+            .begin_write()
+            .map_err(map_transaction_error("Failed to begin write transaction"))?;
+
+        let deleted = {
+            let mut table = write_txn.open_table(NODE_VERSIONS_TABLE).map_err(
+                |e| -> crate::utils::error::Error {
+                    StorageError::io_error(format!("Failed to open node_versions table: {}", e))
+                        .into()
+                },
+            )?;
+
+            match table.remove(id.as_u64()) {
+                Ok(Some(_)) => true,
+                Ok(None) => false,
+                Err(e) => {
+                    return Err(map_storage_error("Failed to delete node version")(e));
+                }
+            }
+        };
+
+        write_txn
+            .commit()
+            .map_err(map_commit_error("Failed to commit"))?;
+
+        Ok(deleted)
+    }
+
+    /// Delete an edge version.
+    pub fn delete_edge_version(&self, id: VersionId) -> Result<bool> {
+        let write_txn = self
+            .db
+            .begin_write()
+            .map_err(map_transaction_error("Failed to begin write transaction"))?;
+
+        let deleted = {
+            let mut table = write_txn.open_table(EDGE_VERSIONS_TABLE).map_err(
+                |e| -> crate::utils::error::Error {
+                    StorageError::io_error(format!("Failed to open edge_versions table: {}", e))
+                        .into()
+                },
+            )?;
+
+            match table.remove(id.as_u64()) {
+                Ok(Some(_)) => true,
+                Ok(None) => false,
+                Err(e) => {
+                    return Err(map_storage_error("Failed to delete edge version")(e));
+                }
+            }
+        };
+
+        write_txn
+            .commit()
+            .map_err(map_commit_error("Failed to commit"))?;
+
+        Ok(deleted)
+    }
+
+    /// Store a batch of node versions.
+    pub fn store_node_versions_batch(&self, versions: &[NodeVersion]) -> Result<()> {
+        self.check_fail_writes()?;
+
+        if versions.is_empty() {
+            return Ok(());
+        }
+
+        let write_txn = self
+            .db
+            .begin_write()
+            .map_err(map_transaction_error("Failed to begin write transaction"))?;
+
+        {
+            let mut table = write_txn.open_table(NODE_VERSIONS_TABLE).map_err(
+                |e| -> crate::utils::error::Error {
+                    StorageError::io_error(format!("Failed to open node_versions table: {}", e))
+                        .into()
+                },
+            )?;
+
+            for version in versions {
+                let encoded = encode_node_version(version);
+                let raw_size = encoded.len();
+                let compressed = self.compress(&encoded)?;
+                let compressed_size = compressed.len();
+
+                table
+                    .insert(version.id.as_u64(), compressed.as_slice())
+                    .map_err(map_storage_error("Failed to store node version"))?;
+
+                self.stats
+                    .node_versions_stored
+                    .fetch_add(1, Ordering::Relaxed);
+                self.stats
+                    .bytes_written_raw
+                    .fetch_add(raw_size as u64, Ordering::Relaxed);
+                self.stats
+                    .bytes_written_compressed
+                    .fetch_add(compressed_size as u64, Ordering::Relaxed);
+            }
+        }
+
+        write_txn
+            .commit()
+            .map_err(map_commit_error("Failed to commit batch"))?;
+
+        Ok(())
+    }
+
+    /// Store a batch of edge versions.
+    pub fn store_edge_versions_batch(&self, versions: &[EdgeVersion]) -> Result<()> {
+        self.check_fail_writes()?;
+
+        if versions.is_empty() {
+            return Ok(());
+        }
+
+        let write_txn = self
+            .db
+            .begin_write()
+            .map_err(map_transaction_error("Failed to begin write transaction"))?;
+
+        {
+            let mut table = write_txn.open_table(EDGE_VERSIONS_TABLE).map_err(
+                |e| -> crate::utils::error::Error {
+                    StorageError::io_error(format!("Failed to open edge_versions table: {}", e))
+                        .into()
+                },
+            )?;
+
+            for version in versions {
+                let encoded = encode_edge_version(version);
+                let raw_size = encoded.len();
+                let compressed = self.compress(&encoded)?;
+                let compressed_size = compressed.len();
+
+                table
+                    .insert(version.id.as_u64(), compressed.as_slice())
+                    .map_err(map_storage_error("Failed to store edge version"))?;
+
+                self.stats
+                    .edge_versions_stored
+                    .fetch_add(1, Ordering::Relaxed);
+                self.stats
+                    .bytes_written_raw
+                    .fetch_add(raw_size as u64, Ordering::Relaxed);
+                self.stats
+                    .bytes_written_compressed
+                    .fetch_add(compressed_size as u64, Ordering::Relaxed);
+            }
+        }
+
+        write_txn
+            .commit()
+            .map_err(map_commit_error("Failed to commit batch"))?;
+
+        Ok(())
+    }
+
+    /// Get usage statistics.
+    pub fn stats(&self) -> ColdStorageStats {
+        self.stats.snapshot()
+    }
+
+    /// Flush to disk.
+    ///
+    /// For Redb, this is a no-op as transactions are durable on commit.
+    pub fn flush(&self) -> Result<()> {
+        // Redb automatically flushes on commit, nothing extra needed
+        Ok(())
+    }
+
+    /// Close the database connection.
+    ///
+    /// This releases the file lock.
+    pub fn close(&self) -> Result<()> {
+        // Redb handles cleanup on drop
+        Ok(())
+    }
+
     /// Store a batch of versions with LSN tracking.
     ///
     /// This operation is atomic - either all versions are stored and the LSN is updated,
@@ -337,6 +901,8 @@ impl RedbColdStorage {
         edges: &[EdgeVersion],
         lsn: LSN,
     ) -> Result<()> {
+        self.check_fail_writes()?;
+
         let write_txn = self
             .db
             .begin_write()
@@ -426,387 +992,397 @@ impl RedbColdStorage {
     }
 }
 
-impl ColdStorage for RedbColdStorage {
-    fn store_node_version(&self, version: &NodeVersion) -> Result<()> {
-        let encoded = encode_node_version(version);
-        let raw_size = encoded.len();
-        let compressed = self.compress(&encoded)?;
-        let compressed_size = compressed.len();
+// Serialization helpers using bitcode
+// ============================================================================
 
-        let write_txn = self
-            .db
-            .begin_write()
-            .map_err(map_transaction_error("Failed to begin write transaction"))?;
+/// Serializable wrapper for NodeVersion.
+#[derive(bitcode::Encode, bitcode::Decode)]
+struct SerializableNodeVersion {
+    id: u64,
+    node_id: u64,
+    temporal_valid_start: i64,
+    temporal_valid_end: i64,
+    temporal_tx_start: i64,
+    temporal_tx_end: i64,
+    label: String,
+    data: SerializableVersionData,
+    next_version: Option<u64>,
+    prev_version: Option<u64>,
+}
 
-        {
-            let mut table = write_txn.open_table(NODE_VERSIONS_TABLE).map_err(
-                |e| -> crate::utils::error::Error {
-                    StorageError::io_error(format!("Failed to open node_versions table: {}", e))
-                        .into()
-                },
-            )?;
+/// Serializable wrapper for EdgeVersion.
+#[derive(bitcode::Encode, bitcode::Decode)]
+struct SerializableEdgeVersion {
+    id: u64,
+    edge_id: u64,
+    temporal_valid_start: i64,
+    temporal_valid_end: i64,
+    temporal_tx_start: i64,
+    temporal_tx_end: i64,
+    label: String,
+    source: u64,
+    target: u64,
+    data: SerializableVersionData,
+    next_version: Option<u64>,
+    prev_version: Option<u64>,
+}
 
-            table
-                .insert(version.id.as_u64(), compressed.as_slice())
-                .map_err(map_storage_error("Failed to store node version"))?;
-        }
+/// Serializable wrapper for VersionData.
+#[derive(bitcode::Encode, bitcode::Decode)]
+enum SerializableVersionData {
+    Anchor {
+        properties: Vec<(String, SerializablePropertyValue)>,
+        vector_snapshot_id: Option<u64>,
+    },
+    Delta {
+        changed: Vec<(String, SerializablePropertyValue)>,
+        removed: Vec<String>,
+    },
+}
 
-        write_txn
-            .commit()
-            .map_err(map_commit_error("Failed to commit"))?;
+/// Serializable property value.
+///
+/// Note: Array is stored as a serialized `Vec<u8>` to avoid recursive type issues with bitcode.
+#[derive(bitcode::Encode, bitcode::Decode)]
+enum SerializablePropertyValue {
+    Null,
+    Bool(bool),
+    Int(i64),
+    Float(f64),
+    String(String),
+    Bytes(Vec<u8>),
+    /// Array is stored as bitcode-encoded bytes to avoid recursive type issues.
+    Array(Vec<u8>),
+    Vector(Vec<f32>),
+    SparseVector {
+        dimension: usize,
+        indices: Vec<u32>,
+        values: Vec<f32>,
+    },
+}
 
-        self.stats
-            .node_versions_stored
-            .fetch_add(1, Ordering::Relaxed);
-        self.stats
-            .bytes_written_raw
-            .fetch_add(raw_size as u64, Ordering::Relaxed);
-        self.stats
-            .bytes_written_compressed
-            .fetch_add(compressed_size as u64, Ordering::Relaxed);
+/// Encode a NodeVersion to bytes for storage.
+///
+/// This function is public for use by all cold storage implementations.
+pub fn encode_node_version(version: &NodeVersion) -> Vec<u8> {
+    use crate::core::interning::GLOBAL_INTERNER;
 
-        Ok(())
+    let serializable = SerializableNodeVersion {
+        id: version.id.as_u64(),
+        node_id: version.node_id.as_u64(),
+        temporal_valid_start: version.temporal.valid_time().start().wallclock(),
+        temporal_valid_end: version.temporal.valid_time().end().wallclock(),
+        temporal_tx_start: version.temporal.transaction_time().start().wallclock(),
+        temporal_tx_end: version.temporal.transaction_time().end().wallclock(),
+        label: GLOBAL_INTERNER
+            .resolve(version.label)
+            .unwrap_or_default()
+            .to_string(),
+        data: encode_version_data(&version.data),
+        next_version: version.next_version.map(|v| v.as_u64()),
+        prev_version: version.prev_version.map(|v| v.as_u64()),
+    };
+
+    bitcode::encode(&serializable)
+}
+
+/// Decode a NodeVersion from bytes.
+///
+/// This function is public for use by all cold storage implementations.
+pub fn decode_node_version(data: &[u8]) -> Result<NodeVersion> {
+    use crate::core::hlc::HybridTimestamp;
+    use crate::core::id::NodeId;
+    use crate::core::interning::GLOBAL_INTERNER;
+    use crate::core::temporal::{BiTemporalInterval, TimeRange};
+
+    let serializable: SerializableNodeVersion = bitcode::decode(data)
+        .map_err(|e| StorageError::corruption(format!("Failed to decode node version: {}", e)))?;
+
+    let valid_time = TimeRange::new(
+        HybridTimestamp::new_unchecked(serializable.temporal_valid_start, 0),
+        HybridTimestamp::new_unchecked(serializable.temporal_valid_end, 0),
+    )
+    .map_err(|e| StorageError::corruption(format!("Invalid valid time range: {}", e)))?;
+    let tx_time = TimeRange::new(
+        HybridTimestamp::new_unchecked(serializable.temporal_tx_start, 0),
+        HybridTimestamp::new_unchecked(serializable.temporal_tx_end, 0),
+    )
+    .map_err(|e| StorageError::corruption(format!("Invalid transaction time range: {}", e)))?;
+
+    Ok(NodeVersion {
+        id: VersionId::new(serializable.id)
+            .map_err(|e| StorageError::corruption(format!("Invalid version ID: {}", e)))?,
+        node_id: NodeId::new(serializable.node_id)
+            .map_err(|e| StorageError::corruption(format!("Invalid node ID: {}", e)))?,
+        temporal: BiTemporalInterval::new(valid_time, tx_time),
+        label: GLOBAL_INTERNER
+            .intern(&serializable.label)
+            .map_err(|e| StorageError::corruption(format!("Failed to intern label: {}", e)))?,
+        data: decode_version_data(serializable.data)?,
+        next_version: serializable
+            .next_version
+            .map(VersionId::new)
+            .transpose()
+            .map_err(|e| StorageError::corruption(format!("Invalid next version ID: {}", e)))?,
+        prev_version: serializable
+            .prev_version
+            .map(VersionId::new)
+            .transpose()
+            .map_err(|e| StorageError::corruption(format!("Invalid prev version ID: {}", e)))?,
+    })
+}
+
+/// Encode an EdgeVersion to bytes for storage.
+///
+/// This function is public for use by all cold storage implementations.
+pub fn encode_edge_version(version: &EdgeVersion) -> Vec<u8> {
+    use crate::core::interning::GLOBAL_INTERNER;
+
+    let serializable = SerializableEdgeVersion {
+        id: version.id.as_u64(),
+        edge_id: version.edge_id.as_u64(),
+        temporal_valid_start: version.temporal.valid_time().start().wallclock(),
+        temporal_valid_end: version.temporal.valid_time().end().wallclock(),
+        temporal_tx_start: version.temporal.transaction_time().start().wallclock(),
+        temporal_tx_end: version.temporal.transaction_time().end().wallclock(),
+        label: GLOBAL_INTERNER
+            .resolve(version.label)
+            .unwrap_or_default()
+            .to_string(),
+        source: version.source.as_u64(),
+        target: version.target.as_u64(),
+        data: encode_version_data(&version.data),
+        next_version: version.next_version.map(|v| v.as_u64()),
+        prev_version: version.prev_version.map(|v| v.as_u64()),
+    };
+
+    bitcode::encode(&serializable)
+}
+
+/// Decode an EdgeVersion from bytes.
+///
+/// This function is public for use by all cold storage implementations.
+pub fn decode_edge_version(data: &[u8]) -> Result<EdgeVersion> {
+    use crate::core::hlc::HybridTimestamp;
+    use crate::core::id::{EdgeId, NodeId};
+    use crate::core::interning::GLOBAL_INTERNER;
+    use crate::core::temporal::{BiTemporalInterval, TimeRange};
+
+    let serializable: SerializableEdgeVersion = bitcode::decode(data)
+        .map_err(|e| StorageError::corruption(format!("Failed to decode edge version: {}", e)))?;
+
+    let valid_time = TimeRange::new(
+        HybridTimestamp::new_unchecked(serializable.temporal_valid_start, 0),
+        HybridTimestamp::new_unchecked(serializable.temporal_valid_end, 0),
+    )
+    .map_err(|e| StorageError::corruption(format!("Invalid valid time range: {}", e)))?;
+    let tx_time = TimeRange::new(
+        HybridTimestamp::new_unchecked(serializable.temporal_tx_start, 0),
+        HybridTimestamp::new_unchecked(serializable.temporal_tx_end, 0),
+    )
+    .map_err(|e| StorageError::corruption(format!("Invalid transaction time range: {}", e)))?;
+
+    Ok(EdgeVersion {
+        id: VersionId::new(serializable.id)
+            .map_err(|e| StorageError::corruption(format!("Invalid version ID: {}", e)))?,
+        edge_id: EdgeId::new(serializable.edge_id)
+            .map_err(|e| StorageError::corruption(format!("Invalid edge ID: {}", e)))?,
+        temporal: BiTemporalInterval::new(valid_time, tx_time),
+        label: GLOBAL_INTERNER
+            .intern(&serializable.label)
+            .map_err(|e| StorageError::corruption(format!("Failed to intern label: {}", e)))?,
+        source: NodeId::new(serializable.source)
+            .map_err(|e| StorageError::corruption(format!("Invalid source ID: {}", e)))?,
+        target: NodeId::new(serializable.target)
+            .map_err(|e| StorageError::corruption(format!("Invalid target ID: {}", e)))?,
+        data: decode_version_data(serializable.data)?,
+        next_version: serializable
+            .next_version
+            .map(VersionId::new)
+            .transpose()
+            .map_err(|e| StorageError::corruption(format!("Invalid next version ID: {}", e)))?,
+        prev_version: serializable
+            .prev_version
+            .map(VersionId::new)
+            .transpose()
+            .map_err(|e| StorageError::corruption(format!("Invalid prev version ID: {}", e)))?,
+    })
+}
+
+fn encode_version_data(data: &crate::storage::version::VersionData) -> SerializableVersionData {
+    use crate::core::interning::GLOBAL_INTERNER;
+    use crate::storage::version::VersionData;
+
+    match data {
+        VersionData::Anchor {
+            properties,
+            vector_snapshot_id,
+        } => SerializableVersionData::Anchor {
+            properties: properties
+                .iter()
+                .map(|(k, v)| {
+                    (
+                        GLOBAL_INTERNER.resolve(*k).unwrap_or_default().to_string(),
+                        encode_property_value(v),
+                    )
+                })
+                .collect(),
+            vector_snapshot_id: vector_snapshot_id.map(|id| id as u64),
+        },
+        VersionData::Delta { delta } => SerializableVersionData::Delta {
+            changed: delta
+                .changed
+                .iter()
+                .map(|(k, v)| {
+                    (
+                        GLOBAL_INTERNER.resolve(*k).unwrap_or_default().to_string(),
+                        encode_property_value(v),
+                    )
+                })
+                .collect(),
+            removed: delta
+                .removed
+                .iter()
+                .map(|k| GLOBAL_INTERNER.resolve(*k).unwrap_or_default().to_string())
+                .collect(),
+        },
     }
+}
 
-    fn get_node_version(&self, id: VersionId) -> Result<Option<NodeVersion>> {
-        self.stats
-            .node_version_reads
-            .fetch_add(1, Ordering::Relaxed);
+fn decode_version_data(
+    data: SerializableVersionData,
+) -> Result<crate::storage::version::VersionData> {
+    use crate::core::interning::GLOBAL_INTERNER;
+    use crate::core::property::PropertyMapBuilder;
+    use crate::storage::version::{PropertyDelta, VersionData};
 
-        let read_txn = self
-            .db
-            .begin_read()
-            .map_err(map_transaction_error("Failed to begin read transaction"))?;
-
-        let table = read_txn.open_table(NODE_VERSIONS_TABLE).map_err(
-            |e| -> crate::utils::error::Error {
-                StorageError::io_error(format!("Failed to open node_versions table: {}", e)).into()
-            },
-        )?;
-
-        match table.get(id.as_u64()) {
-            Ok(Some(value)) => {
-                let compressed: &[u8] = value.value();
-                self.stats
-                    .bytes_read_compressed
-                    .fetch_add(compressed.len() as u64, Ordering::Relaxed);
-
-                let decompressed = self.decompress(compressed)?;
-                self.stats
-                    .bytes_read_decompressed
-                    .fetch_add(decompressed.len() as u64, Ordering::Relaxed);
-
-                let version = decode_node_version(&decompressed)?;
-                Ok(Some(version))
+    match data {
+        SerializableVersionData::Anchor {
+            properties,
+            vector_snapshot_id,
+        } => {
+            let mut builder = PropertyMapBuilder::new();
+            for (key, value) in properties {
+                builder = builder.insert(&key, decode_property_value(value)?);
             }
-            Ok(None) => Ok(None),
-            Err(e) => {
-                Err(StorageError::io_error(format!("Failed to read node version: {}", e)).into())
+            Ok(VersionData::Anchor {
+                properties: builder.build(),
+                vector_snapshot_id: vector_snapshot_id.map(|id| id as usize),
+            })
+        }
+        SerializableVersionData::Delta { changed, removed } => {
+            let mut delta = PropertyDelta::new();
+            for (key, value) in changed {
+                let interned_key = GLOBAL_INTERNER.intern(&key).map_err(|e| {
+                    StorageError::corruption(format!("Failed to intern key: {}", e))
+                })?;
+                delta
+                    .changed
+                    .insert(interned_key, decode_property_value(value)?);
             }
+            for key in removed {
+                let interned_key = GLOBAL_INTERNER.intern(&key).map_err(|e| {
+                    StorageError::corruption(format!("Failed to intern key: {}", e))
+                })?;
+                delta.removed.insert(interned_key);
+            }
+            Ok(VersionData::Delta { delta })
         }
     }
+}
 
-    fn store_edge_version(&self, version: &EdgeVersion) -> Result<()> {
-        let encoded = encode_edge_version(version);
-        let raw_size = encoded.len();
-        let compressed = self.compress(&encoded)?;
-        let compressed_size = compressed.len();
+/// Simple array element for encoding (non-recursive).
+#[derive(bitcode::Encode, bitcode::Decode)]
+enum SimpleArrayElement {
+    Null,
+    Bool(bool),
+    Int(i64),
+    Float(f64),
+    String(String),
+    Bytes(Vec<u8>),
+    Vector(Vec<f32>),
+}
 
-        let write_txn = self
-            .db
-            .begin_write()
-            .map_err(map_transaction_error("Failed to begin write transaction"))?;
+fn encode_property_value(
+    value: &crate::core::property::PropertyValue,
+) -> SerializablePropertyValue {
+    use crate::core::property::PropertyValue;
 
-        {
-            let mut table = write_txn.open_table(EDGE_VERSIONS_TABLE).map_err(
-                |e| -> crate::utils::error::Error {
-                    StorageError::io_error(format!("Failed to open edge_versions table: {}", e))
-                        .into()
-                },
-            )?;
-
-            table
-                .insert(version.id.as_u64(), compressed.as_slice())
-                .map_err(map_storage_error("Failed to store edge version"))?;
+    match value {
+        PropertyValue::Null => SerializablePropertyValue::Null,
+        PropertyValue::Bool(b) => SerializablePropertyValue::Bool(*b),
+        PropertyValue::Int(i) => SerializablePropertyValue::Int(*i),
+        PropertyValue::Float(f) => SerializablePropertyValue::Float(*f),
+        PropertyValue::String(s) => SerializablePropertyValue::String(s.to_string()),
+        PropertyValue::Bytes(b) => SerializablePropertyValue::Bytes(b.to_vec()),
+        PropertyValue::Array(arr) => {
+            // Encode array elements as simple types (no nested arrays supported)
+            let elements: Vec<SimpleArrayElement> = arr
+                .iter()
+                .map(|v| match v {
+                    PropertyValue::Null => SimpleArrayElement::Null,
+                    PropertyValue::Bool(b) => SimpleArrayElement::Bool(*b),
+                    PropertyValue::Int(i) => SimpleArrayElement::Int(*i),
+                    PropertyValue::Float(f) => SimpleArrayElement::Float(*f),
+                    PropertyValue::String(s) => SimpleArrayElement::String(s.to_string()),
+                    PropertyValue::Bytes(b) => SimpleArrayElement::Bytes(b.to_vec()),
+                    PropertyValue::Vector(v) => SimpleArrayElement::Vector(v.to_vec()),
+                    // Nested arrays and sparse vectors are not supported in cold storage arrays
+                    _ => SimpleArrayElement::Null,
+                })
+                .collect();
+            SerializablePropertyValue::Array(bitcode::encode(&elements))
         }
-
-        write_txn
-            .commit()
-            .map_err(map_commit_error("Failed to commit"))?;
-
-        self.stats
-            .edge_versions_stored
-            .fetch_add(1, Ordering::Relaxed);
-        self.stats
-            .bytes_written_raw
-            .fetch_add(raw_size as u64, Ordering::Relaxed);
-        self.stats
-            .bytes_written_compressed
-            .fetch_add(compressed_size as u64, Ordering::Relaxed);
-
-        Ok(())
+        PropertyValue::Vector(v) => SerializablePropertyValue::Vector(v.to_vec()),
+        PropertyValue::SparseVector(sv) => SerializablePropertyValue::SparseVector {
+            dimension: sv.dimension(),
+            indices: sv.indices().to_vec(),
+            values: sv.values().to_vec(),
+        },
     }
+}
 
-    fn get_edge_version(&self, id: VersionId) -> Result<Option<EdgeVersion>> {
-        self.stats
-            .edge_version_reads
-            .fetch_add(1, Ordering::Relaxed);
+fn decode_property_value(
+    value: SerializablePropertyValue,
+) -> Result<crate::core::property::PropertyValue> {
+    use crate::core::property::PropertyValue;
+    use crate::core::vector::SparseVec;
 
-        let read_txn = self
-            .db
-            .begin_read()
-            .map_err(map_transaction_error("Failed to begin read transaction"))?;
-
-        let table = read_txn.open_table(EDGE_VERSIONS_TABLE).map_err(
-            |e| -> crate::utils::error::Error {
-                StorageError::io_error(format!("Failed to open edge_versions table: {}", e)).into()
-            },
-        )?;
-
-        match table.get(id.as_u64()) {
-            Ok(Some(value)) => {
-                let compressed: &[u8] = value.value();
-                self.stats
-                    .bytes_read_compressed
-                    .fetch_add(compressed.len() as u64, Ordering::Relaxed);
-
-                let decompressed = self.decompress(compressed)?;
-                self.stats
-                    .bytes_read_decompressed
-                    .fetch_add(decompressed.len() as u64, Ordering::Relaxed);
-
-                let version = decode_edge_version(&decompressed)?;
-                Ok(Some(version))
-            }
-            Ok(None) => Ok(None),
-            Err(e) => {
-                Err(StorageError::io_error(format!("Failed to read edge version: {}", e)).into())
-            }
+    Ok(match value {
+        SerializablePropertyValue::Null => PropertyValue::Null,
+        SerializablePropertyValue::Bool(b) => PropertyValue::Bool(b),
+        SerializablePropertyValue::Int(i) => PropertyValue::Int(i),
+        SerializablePropertyValue::Float(f) => PropertyValue::Float(f),
+        SerializablePropertyValue::String(s) => PropertyValue::string(&s),
+        SerializablePropertyValue::Bytes(b) => PropertyValue::bytes(&b),
+        SerializablePropertyValue::Array(encoded) => {
+            let elements: Vec<SimpleArrayElement> = bitcode::decode(&encoded)
+                .map_err(|e| StorageError::corruption(format!("Failed to decode array: {}", e)))?;
+            let values: Vec<PropertyValue> = elements
+                .into_iter()
+                .map(|e| match e {
+                    SimpleArrayElement::Null => PropertyValue::Null,
+                    SimpleArrayElement::Bool(b) => PropertyValue::Bool(b),
+                    SimpleArrayElement::Int(i) => PropertyValue::Int(i),
+                    SimpleArrayElement::Float(f) => PropertyValue::Float(f),
+                    SimpleArrayElement::String(s) => PropertyValue::string(&s),
+                    SimpleArrayElement::Bytes(b) => PropertyValue::bytes(&b),
+                    SimpleArrayElement::Vector(v) => PropertyValue::vector(&v),
+                })
+                .collect();
+            PropertyValue::array(values)
         }
-    }
-
-    fn contains_node_version(&self, id: VersionId) -> Result<bool> {
-        let read_txn = self
-            .db
-            .begin_read()
-            .map_err(map_transaction_error("Failed to begin read transaction"))?;
-
-        let table = read_txn.open_table(NODE_VERSIONS_TABLE).map_err(
-            |e| -> crate::utils::error::Error {
-                StorageError::io_error(format!("Failed to open node_versions table: {}", e)).into()
-            },
-        )?;
-
-        match table.get(id.as_u64()) {
-            Ok(Some(_)) => Ok(true),
-            Ok(None) => Ok(false),
-            Err(e) => {
-                Err(StorageError::io_error(format!("Failed to check node version: {}", e)).into())
-            }
+        SerializablePropertyValue::Vector(v) => PropertyValue::vector(&v),
+        SerializablePropertyValue::SparseVector {
+            dimension,
+            indices,
+            values,
+        } => {
+            let sparse = SparseVec::new(indices, values, dimension as u32)
+                .map_err(|e| StorageError::corruption(format!("Invalid sparse vector: {}", e)))?;
+            PropertyValue::sparse_vector(sparse)
         }
-    }
-
-    fn contains_edge_version(&self, id: VersionId) -> Result<bool> {
-        let read_txn = self
-            .db
-            .begin_read()
-            .map_err(map_transaction_error("Failed to begin read transaction"))?;
-
-        let table = read_txn.open_table(EDGE_VERSIONS_TABLE).map_err(
-            |e| -> crate::utils::error::Error {
-                StorageError::io_error(format!("Failed to open edge_versions table: {}", e)).into()
-            },
-        )?;
-
-        match table.get(id.as_u64()) {
-            Ok(Some(_)) => Ok(true),
-            Ok(None) => Ok(false),
-            Err(e) => {
-                Err(StorageError::io_error(format!("Failed to check edge version: {}", e)).into())
-            }
-        }
-    }
-
-    fn delete_node_version(&self, id: VersionId) -> Result<bool> {
-        let write_txn = self
-            .db
-            .begin_write()
-            .map_err(map_transaction_error("Failed to begin write transaction"))?;
-
-        let deleted = {
-            let mut table = write_txn.open_table(NODE_VERSIONS_TABLE).map_err(
-                |e| -> crate::utils::error::Error {
-                    StorageError::io_error(format!("Failed to open node_versions table: {}", e))
-                        .into()
-                },
-            )?;
-
-            match table.remove(id.as_u64()) {
-                Ok(Some(_)) => true,
-                Ok(None) => false,
-                Err(e) => {
-                    return Err(map_storage_error("Failed to delete node version")(e));
-                }
-            }
-        };
-
-        write_txn
-            .commit()
-            .map_err(map_commit_error("Failed to commit"))?;
-
-        Ok(deleted)
-    }
-
-    fn delete_edge_version(&self, id: VersionId) -> Result<bool> {
-        let write_txn = self
-            .db
-            .begin_write()
-            .map_err(map_transaction_error("Failed to begin write transaction"))?;
-
-        let deleted = {
-            let mut table = write_txn.open_table(EDGE_VERSIONS_TABLE).map_err(
-                |e| -> crate::utils::error::Error {
-                    StorageError::io_error(format!("Failed to open edge_versions table: {}", e))
-                        .into()
-                },
-            )?;
-
-            match table.remove(id.as_u64()) {
-                Ok(Some(_)) => true,
-                Ok(None) => false,
-                Err(e) => {
-                    return Err(map_storage_error("Failed to delete edge version")(e));
-                }
-            }
-        };
-
-        write_txn
-            .commit()
-            .map_err(map_commit_error("Failed to commit"))?;
-
-        Ok(deleted)
-    }
-
-    fn store_node_versions_batch(&self, versions: &[NodeVersion]) -> Result<()> {
-        if versions.is_empty() {
-            return Ok(());
-        }
-
-        let write_txn = self
-            .db
-            .begin_write()
-            .map_err(map_transaction_error("Failed to begin write transaction"))?;
-
-        {
-            let mut table = write_txn.open_table(NODE_VERSIONS_TABLE).map_err(
-                |e| -> crate::utils::error::Error {
-                    StorageError::io_error(format!("Failed to open node_versions table: {}", e))
-                        .into()
-                },
-            )?;
-
-            for version in versions {
-                let encoded = encode_node_version(version);
-                let raw_size = encoded.len();
-                let compressed = self.compress(&encoded)?;
-                let compressed_size = compressed.len();
-
-                table
-                    .insert(version.id.as_u64(), compressed.as_slice())
-                    .map_err(map_storage_error("Failed to store node version"))?;
-
-                self.stats
-                    .node_versions_stored
-                    .fetch_add(1, Ordering::Relaxed);
-                self.stats
-                    .bytes_written_raw
-                    .fetch_add(raw_size as u64, Ordering::Relaxed);
-                self.stats
-                    .bytes_written_compressed
-                    .fetch_add(compressed_size as u64, Ordering::Relaxed);
-            }
-        }
-
-        write_txn
-            .commit()
-            .map_err(map_commit_error("Failed to commit batch"))?;
-
-        Ok(())
-    }
-
-    fn store_edge_versions_batch(&self, versions: &[EdgeVersion]) -> Result<()> {
-        if versions.is_empty() {
-            return Ok(());
-        }
-
-        let write_txn = self
-            .db
-            .begin_write()
-            .map_err(map_transaction_error("Failed to begin write transaction"))?;
-
-        {
-            let mut table = write_txn.open_table(EDGE_VERSIONS_TABLE).map_err(
-                |e| -> crate::utils::error::Error {
-                    StorageError::io_error(format!("Failed to open edge_versions table: {}", e))
-                        .into()
-                },
-            )?;
-
-            for version in versions {
-                let encoded = encode_edge_version(version);
-                let raw_size = encoded.len();
-                let compressed = self.compress(&encoded)?;
-                let compressed_size = compressed.len();
-
-                table
-                    .insert(version.id.as_u64(), compressed.as_slice())
-                    .map_err(map_storage_error("Failed to store edge version"))?;
-
-                self.stats
-                    .edge_versions_stored
-                    .fetch_add(1, Ordering::Relaxed);
-                self.stats
-                    .bytes_written_raw
-                    .fetch_add(raw_size as u64, Ordering::Relaxed);
-                self.stats
-                    .bytes_written_compressed
-                    .fetch_add(compressed_size as u64, Ordering::Relaxed);
-            }
-        }
-
-        write_txn
-            .commit()
-            .map_err(map_commit_error("Failed to commit batch"))?;
-
-        Ok(())
-    }
-
-    fn stats(&self) -> ColdStorageStats {
-        self.stats.snapshot()
-    }
-
-    fn flush(&self) -> Result<()> {
-        // Redb automatically flushes on commit, nothing extra needed
-        Ok(())
-    }
-
-    fn close(&self) -> Result<()> {
-        // Redb handles cleanup on drop
-        Ok(())
-    }
-
-    fn get_flushed_lsn(&self) -> Result<Option<LSN>> {
-        RedbColdStorage::get_flushed_lsn(self)
-    }
-
-    fn store_batch_with_lsn(
-        &self,
-        nodes: &[NodeVersion],
-        edges: &[EdgeVersion],
-        lsn: LSN,
-    ) -> Result<()> {
-        RedbColdStorage::store_batch_with_lsn(self, nodes, edges, lsn)
-    }
+    })
 }
 
 #[cfg(test)]
@@ -1756,32 +2332,6 @@ mod tests {
     }
 
     #[test]
-    fn test_trait_delegation_methods() {
-        // Test that trait methods properly delegate to inherent methods
-        let temp_dir = tempfile::tempdir().unwrap();
-        let db_path = temp_dir.path().join("test.redb");
-        let storage = RedbColdStorage::with_default_config(&db_path).unwrap();
-
-        // Test trait get_flushed_lsn delegates to inherent method
-        use crate::storage::ColdStorage;
-        assert_eq!(ColdStorage::get_flushed_lsn(&storage).unwrap(), None);
-
-        // Store with LSN using trait method
-        let nodes = vec![create_test_node_version(1)];
-        ColdStorage::store_batch_with_lsn(&storage, &nodes, &[], LSN(300)).unwrap();
-
-        // Verify using both trait and inherent methods
-        assert_eq!(
-            ColdStorage::get_flushed_lsn(&storage).unwrap(),
-            Some(LSN(300))
-        );
-        assert_eq!(
-            RedbColdStorage::get_flushed_lsn(&storage).unwrap(),
-            Some(LSN(300))
-        );
-    }
-
-    #[test]
     fn test_config_cache_size() {
         let config = RedbConfig::new().cache_size_bytes(4096);
         assert_eq!(config.cache_size_bytes, 4096);
@@ -2647,5 +3197,20 @@ mod tests {
                     .unwrap()
             );
         }
+    }
+
+    #[test]
+    fn test_fault_injection() {
+        let temp_dir = tempfile::tempdir().unwrap();
+        let db_path = temp_dir.path().join("test.redb");
+        let storage = RedbColdStorage::with_default_config(&db_path).unwrap();
+
+        storage.set_fail_writes(true);
+
+        let node = create_test_node_version(1);
+        let result = storage.store_node_version(&node);
+
+        assert!(result.is_err());
+        assert!(storage.was_write_attempted());
     }
 }
